+++ src/mocks/browser.ts
@@ -1,4 +1,5 @@
 import { setupWorker } from "msw";
 import { handlers } from "./handlers";
+import { chartHandlers } from "./handlers/chartHandlers";
 
-export const worker = setupWorker(...handlers);
+export const worker = setupWorker(...handlers, ...chartHandlers);
