+++ src/mocks/handlers/chartHandlers.ts
@@ -1,8 +1,8 @@
 import { rest } from "msw";
 import { STATISTIC_DATA } from "../database/chart";
 
-export const handlers = [
-  rest.get("/result", async (req, res, ctx) => {
+export const chartHandlers = [
+  rest.get("/statistics", async (req, res, ctx) => {
     return res(
       ctx.json({
         STATISTIC_DATA,
